# Up to 10x speedup when parsing attribute injection