One reason why parsing attribute injection is slow is because the same topology strings keep getting parsed again and again. This PR fixes this by caching previously parsed topology strings.

I've seen an up to 10x improvement where what took ~42s now takes ~4.8s and what took 100s now takes ~10s.